// edTechSpark_frontend_v18/edTechSpark_frontend_v18/edtechspark/src/app/models/courselesson.ts
export class CourseLesson {
    id: number | any;
    lessonName: string | undefined;
    videoPath: string | undefined;
    contentPath: string | undefined;
    isPreview: boolean | undefined;
    isActive: boolean | undefined;
    duration: string | undefined;
    sequence: number | undefined;
    courseTopicId: number | undefined;
    createdDate: Date | undefined;
    updatedDate: Date | undefined;
}